+++ docs/model-persistence.md
@@ -4,16 +4,16 @@ Model persistence is the ability to save and subsequently load a learner's state
 
 ## Serialization
 
-Serialization occurs in between saving and loading a model and can be thought of as packaging the model's parameters. The data can be in a lightweight format such as with PHP's [Native](serializers/native.md) serializer or in a robust format such as [RBXV2](serializers/rbx-v2.md). In the this example, we'll demonstrate how to encode a Persistable learner using the robust RBXV2 format, save the encoding with a [Persister](persisters/api.md), and then how to deserialize the encoding.
+Serialization occurs in between saving and loading a model and can be thought of as packaging the model's parameters. The data can be in a lightweight format such as with PHP's [Native](serializers/native.md) serializer or in a robust format such as [RBX](serializers/rbx-v1.md). In the this example, we'll demonstrate how to encode a Persistable learner using the RBX format, save the encoding with a [Persister](persisters/api.md), and then how to deserialize the encoding.
 
 ```php
 use Rubix\ML\Classifiers\RandomForest;
-use Rubix\ML\Serializers\RBXV2;
+use Rubix\ML\Serializers\RBX;
 use Rubix\ML\Persisters\Filesystem;
 
 $estimator = new RandomForest(100);
 
-$serializer = new RBXV2();
+$serializer = new RBX();
 
 $encoding = $serializer->serialize($estimator);
 
@@ -52,12 +52,12 @@ In addition to Learners, the persistence subsystem can be used to individually s
 
 ```php
 use Rubix\ML\Transformers\OneHotEncoder;
-use Rubix\ML\Serializers\RBXV2;
+use Rubix\ML\Serializers\RBX;
 use Rubix\ML\Persisters\Filesystem;
 
 $transformer = new OneHotEncoder();
 
-$serializer = new RBXV2();
+$serializer = new RBX();
 
 $transformer->fit($dataset);
 
@@ -67,10 +67,10 @@ $serializer->serialize($transformer)->saveTo(new Filesystem('example.rbx'));
 Then, to load the transformer in another process call the `deserialize()` method on the encoding returned by the persister's `load()` method.
 
 ```php
-use Rubix\ML\Serializers\RBXV2;
+use Rubix\ML\Serializers\RBX;
 use Rubix\ML\Persisters\Filesystem;
 
-$transformer = $persister->load()->deserializeWith(new RBXV2());
+$transformer = $persister->load()->deserializeWith(new RBX());
 ```
 
 ## Caveats
