+++ docs/persistent-model.md
@@ -14,7 +14,7 @@ The Persistent Model meta-estimator wraps a [Persistable](persistable.md) learne
 | --- | --- | --- | --- | --- |
 | 1 | base | | Persistable | The persistable base learner. |
 | 2 | persister | | Persister | The persister used to interface with the storage system. |
-| 3 | serializer | RBXV2 | Serializer | The object serializer. |
+| 3 | serializer | RBX | Serializer | The object serializer. |
 
 ## Examples
 
@@ -22,9 +22,9 @@ The Persistent Model meta-estimator wraps a [Persistable](persistable.md) learne
 use Rubix\ML\PersistentModel;
 use Rubix\ML\Clusterers\KMeans;
 use Rubix\ML\Persisters\Filesystem;
-use Rubix\ML\Serializers\RBXV2;
+use Rubix\ML\Serializers\RBX;
 
-$estimator = new PersistentModel(new KMeans(10), new Filesystem('example.model'), new RBXV2());
+$estimator = new PersistentModel(new KMeans(10), new Filesystem('example.model'), new RBX());
 ```
 
 ## Additional Methods
@@ -38,9 +38,9 @@ public static load(Persister $persister, ?Serializer $serializer = null) : self
 ```php
 use Rubix\ML\PersistentModel;
 use Rubix\ML\Persisters\Filesystem;
-use Rubix\ML\Serializers\RBXV2;
+use Rubix\ML\Serializers\RBX;
 
-$estimator = PersistentModel::load(new Filesystem('example.model'), new RBXV2());
+$estimator = PersistentModel::load(new Filesystem('example.model'), new RBX());
 ```
 
 Save the model to storage:
